backend/requirements.txt
azure-core==1.30.1
azure-cosmos==4.7.0
azure-storage-blob==12.20.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.5.0
huggingface-hub==0.23.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pycparser==2.22
PyJWT==2.8.0
pymongo==4.7.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
six==1.16.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3